extract deltas writer
