+++ src-tauri/src/deltas/mod.rs
@@ -3,9 +3,11 @@ mod delta;
 mod document;
 mod operations;
 mod reader;
+mod writer;
 
 pub use database::Database;
 pub use delta::Delta;
 pub use document::Document;
 pub use operations::Operation;
 pub use reader::DeltasReader as Reader;
+pub use writer::DeltasWriter as Writer;
